adding GC content, microexon analysis of gencode, vastdb to pipeline

+++ nullomers/bin/pipeline/7-phylop_expand-array.py
@@ -1,5 +1,4 @@
 import glob
-from joblib import Parallel, delayed
 import os
 import pybedtools as pbt
 import subprocess
@@ -39,7 +38,6 @@ config_name = os.path.join(os.getcwd(), config_tag)
 
 config, configname = crw.read_config(config_name)
 
-
 # select config variables
 ANNOT = config["GENCODE"]["ANNOT"]  
 DATA_PATH = config["DATA"]["PATH"]
@@ -77,52 +75,6 @@ def getArrayMatch(arrayfile, array_ind):
     return RETURN_BED
     
 
-def expand_bedcoor(bedfile, nbases):
-    """
-    write expanded bedcoordinates file. 
-    
-    inputs
-        bedfile (str) - full path to bed file to expand
-        nbases (int) - number of bases to expand by IN EACH DIRECTION
-        
-    method
-        
-        1. build outfile str
-        2. make pybedtool object (if outfile does not exist, else skip)
-        3. use bedtools slop to expand bed coordinates, save output
-        4. unit test expansion
-        
-    returns
-    
-        outbed (str) - full path to expanded coordinate bedfile
-    """
-    
-    #1
-    out = bedfile.strip(".bed") + f"-EXPANDED-{nbases}.bed"
-    
-    if os.path.exists(out) is False:
-        #2
-        bed = pbt.BedTool(bedfile)
-
-        #3
-        expanded = bed.slop(genome="hg38", b=nbases, output=out)
-        
-    else:
-        print("expanded", nbases, " in either direction, already")
-
-    #4
-    with open(bedfile, "r") as file:
-        for line in file:
-            print("original", line)
-            break  # only print the first original line in the file
-            
-    with open(out, "r") as ex:
-        for line in ex:
-            print("expanded", line) 
-            break  # only print he first expanded line in the file
-    
-    return out
-
 # ## parse bigwig for bed coordinates
 
 
@@ -147,14 +99,11 @@ def getBwBedLineSummary(bw, bed_line_list, nbases, outfile, src):
         3. make new line (str) w/ bedcoordinates, id, phylop results
         4. write bed coordinates and phylop to the outfile
         
-        
     """
     
     #1
     chr_, start_, end_, nid = bed_line_list
     nsteps = int(nbases)*2
-    
-    # TO DO: insert a check for existing file, number of lines in file. Concerned about parallelized runs and writing to wrong coordinates. 
     
     #2              
     cmd = f"{src} {bw} {chr_} {str(start_)} {str(end_)} {str(nsteps)}"
@@ -173,30 +122,42 @@ def getBwBedLineSummary(bw, bed_line_list, nbases, outfile, src):
 
 # get summary stats for phylop scores
 
-def getSummaryStats(phylop_file):
-    
+def q025(x):
+    return x.quantile(0.025)
+
+def q975(x):
+    return x.quantile(0.975)
+
+
+def getSummaryStats(file, value_col):
     """
-    return summary stats for each position in a phylop window
+    return summary stats for each position in a bed element
     
     input
-        phylop_file (str) - path to phylop bw extracted file from above run
+        file (str) - path to extracted bw vector file from above run
+        value_col (str) - name of column to calculate values on. 
         
     method
         1. make a pandas dataframe out of the dictionary
             
             col 0-3 = CHR, START, STOP, Nullomer ID
-            col 4-1003 = each bp position phylop score flanking nullomer
+            col 4-1003 = each bp position score flanking nullomer
             
-            col 503 = nullomer. 
+            col 503 = pre-nullomer locus.
+            
+            1.1 Formatting
+                - drop_duplicates() 
+                - dropna()
 
-        2. filter
-            2.1 remove any lines w/ no data 
-            2.2 replace "n/a" str value
-        
-        3. turn dataframe into long form data, where 1 column is the position, and another column is the phylop value
+    
+        2. turn dataframe into long form data, where 1 column is the position, and another column is the value
             index on CHR, START, STOP, Nullomer ID
+            
+            - DROP NAN values. 
+            
+        3. per position, bootstrap confidence intervals. 
         
-        4. compute summary stats per position across nullomer sequence for phylop scores.
+        4. compute summary stats per position across nullomer sequence for linsight scores.
             4.1 also correct for
         
         5. write summary stats to outfile
@@ -208,34 +169,44 @@ def getSummaryStats(phylop_file):
             
         
     """
-    out = phylop_file.strip(".txt") + "-SUMMARY_STATS.txt"
+    out = file.strip(".txt") + "-SUMMARY_STATS.txt"
+    zipped = file + ".gz" 
 
-    #1 make a pandas dataframe
-    df = pd.read_csv(phylop_file, sep = '\t', header = None, low_memory = False)
-    
+    if os.path.exists(out) is False:
+        
+        # if re-running stats
+        if os.path.exists(zipped) is True:
+            zippery.unzip_file(zipped)
+            
+        
+        #1 make a pandas dataframe
+        df = pd.read_csv(file, sep = '\t', header = None, low_memory = False).drop_duplicates().dropna() #1.1
 
-    #2
-    df = df.loc[df[4] != "no"] ## there are some cases where there is no data. 
-    df = df.replace("n/a", None)
-    print(df.shape)
-    print(df.head())
+        #2
 
-    
-    #3
-    melted = pd.melt(df, id_vars= df.columns[:4], var_name="pos", value_name="phylop")
-    print(melted.head())
-    
-    melted["phylop"] = melted["phylop"].astype(float)  # change data type
-    
-    #4
-    summary_stats = melted.groupby("pos")["phylop"].describe().reset_index()
-    
-    summary_stats["pos"] = summary_stats["pos"] - 4 # correct bp position 
-    
-    #5 
-    summary_stats.to_csv(out, sep ='\t', index=False)
-    
-    zippery.rezip_file(phylop_file)
+        melted = pd.melt(df, id_vars= df.columns[:4], var_name="pos", value_name=value_col)
+        print(melted.shape, melted.head())
+
+        melted[value_col] = melted[value_col].astype(float)  # change data type
+        
+
+        f = {value_col: [q025, q975]}
+        ci = melted.groupby('pos').agg(f).reset_index()
+
+        #4
+        summary_stats = melted.groupby("pos")[value_col].describe().reset_index()
+        summary_stats= summary_stats.join(ci)
+
+        print(list(summary_stats))
+        
+
+        summary_stats["index"] = summary_stats["index"] - 4 # correct bp position 
+                            
+
+        #5 
+        summary_stats.to_csv(out, sep ='\t', index=False)
+
+        zippery.rezip_file(file)
     
     return out
 
@@ -245,7 +216,6 @@ def getSummaryStats(phylop_file):
 
 def main(argv):
     
-
     
     # combined function that expands mutations and extracts phylop scores
     #out = expandAndBwExtract(BED, NBASES, PHYLOP_BW, PHYLOP_PATH)
@@ -255,7 +225,7 @@ def main(argv):
     inputs
         BED (str) - bed file name w/ full path from ARRAY
         PHYLOP_PATH (str) - full path to output file directory
-        NBASES (int) - number of bases to expand on each side
+
         PHYLOP_BW (str) - full path to phylop bigWig file. 
         
         
@@ -263,9 +233,9 @@ def main(argv):
     0. makes an outfile. 
         0.1 if scores have not been extracted already
         0.2 if scores have been and the file is zipped
-    1. takes locus and expands the bed coordinates of the locus. 
-    2. extracts the phylop base scores for the expanded locus, writing the scores as a vector
-    3. write summary stats of file. (Summary stats at each bp)
+    
+    1. extracts the phylop base scores for the expanded locus, writing the scores as a vector
+    2. write summary stats of file. (Summary stats at each bp)
 
 
     return 
@@ -273,34 +243,28 @@ def main(argv):
     """
     
     BED = getArrayMatch(ARRAY, ARRAY_IND)
-    
-    
 
     #0 make the outfile
     name = (BED.split("/")[-1]).strip(".bed") + "-phylop.txt"
     outfile = os.path.join(PHYLOP_PATH, name)
     
-
-    #1 expand bed coordinates
-    print("expanding", BED, "n bases", NBASES)
-    
-    outex = expand_bedcoor(BED, int(NBASES))
-    
-    print("EXPANDED", outex)
-
-    #2 get phylop values from bigwig and write to the outfile. 
-    with open(outex, "r") as ex:
-    
-        print("reading expanded file", outex)
+    if os.path.exists(outfile +".gz") is False and os.path.exists(outfile) is False:
         
-        for line in ex.readlines():
-            bedlist = (line.strip("\n")).split("\t")
-            getBwBedLineSummary(PHYLOP_BW, bedlist, NBASES, outfile, SRC_BW)
+        
 
-    print("wrote phylop scores to", outfile)
+        #1 get phylop values from bigwig and write to the outfile. 
+        with open(BED, "r") as ex:
 
-    # 3
-    summary_outfile = getSummaryStats(outfile)
+            print("reading expanded file", BED)
+
+            for line in ex.readlines():
+                bedlist = (line.strip("\n")).split("\t")
+                getBwBedLineSummary(PHYLOP_BW, bedlist, NBASES, outfile, SRC_BW)
+
+        print("wrote phylop scores to", outfile)
+
+    # 2
+    summary_outfile = getSummaryStats(outfile, "phylop")
 
     print(summary_outfile)
